feat: rename anything that sounds OOey and remove old OOP residue
